Added IR generation guards for exact-must-aliasing and a small optimization to use that information.
Besides the malloc-based no-alias checks we can now also generate guards that check for pointer equality.
If we have proven that two pointers are equal within a loop we can then replace all uses of the one with the other, hoping that LLVM can then do some redundant code elimination or code hoisting.